Commented out ConnectDB.java, created HotelReservationSystem.java and implemented ConnectDB as method
